requirements/base.txt
Django==1.11.29
djangorestframework==3.9.4
psycopg2==2.8.6
pytz==2022.7
